Automatically redirect to originally requested path after login
When a user tries to access the site and they are not logged in,
they will be prompted to login. This commit introduces functionality
to automatically redirect the user to the URI they originally requested
after a successful login. The URI data will persist alongside the
login info, even when an unsuccesfull login attempt is performed.